requirements.txt
tensorflow>=2.16
numpy
pandas
scikit-learn
matplotlib
pyyaml
